// src/models/user.model.ts
export class UserSchema {
  id?: number;
  name?: string;
  lastname?: string;
  email?: string;
  date_birthday?: Date;
  description?: string;
  phone?: string;
  // constructor(
  //   id: number,
  //   name: string,
  //   lastname: string,
  //   email: string,
  //   dateBirthday: Date,
  //   description: string,
  //   phone: string
  // ) {
  //   this.name = name;
  //   this.lastname = lastname;
  //   this.email = email;
  //   this.date_birthday = dateBirthday;
  //   this.phone = phone;
  //   this.description = description;
  // }
}